+++ src/main.js
@@ -1,5 +1,6 @@
 import { createApp } from 'vue'
 import './style/index.scss'
 import App from './App.vue'
+import router from './router'
 
-createApp(App).mount('#app')
+createApp(App).use(router).mount('#app')
